eventHash in bundle from float to int

+++ app/src/androidTest/java/ch/epfl/sdp/ui/main/swipe/SwipeFragmentTest.java
@@ -199,7 +199,7 @@ public class SwipeFragmentTest {
         mockSetup();
 
         Bundle bundle = new Bundle();
-        bundle.putFloat("eventHash", eventTest2.hashCode());
+        bundle.putInt("eventHash", eventTest2.hashCode());
 
         mScenario= FragmentScenario.launchInContainer(
                 SwipeFragment.class,
